+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 CHANGELOG
 
 ## vNext
+- Added `Primitives.ImmediateProgress<T>`, a synchronous `System.IProgress<T>`, unlike `System.Progress<T>`.
+- Added `Primitives.NullProgress<T>`, a Null Object implementation of `System.IProgress<T>` as Singleton.
 
 ## v0.1.0 (2018-09-18)
-- Added Mathematics.Parity checking whether an integer is even or odd.
+- Added `Mathematics.Parity` checking whether an integer is even or odd.